Fix build linux64 module with DEBUG enabled.

+++ sys/compat/linux/linux_misc.c
@@ -901,7 +901,8 @@ linux_mknod(struct thread *td, struct linux_mknod_args *args)
 
 #ifdef DEBUG
 	if (ldebug(mknod))
-		printf(ARGS(mknod, "%s, %d, %d"), path, args->mode, args->dev);
+		printf(ARGS(mknod, "%s, %d, %jd"), path, args->mode,
+		    (intmax_t)args->dev);
 #endif
 
 	switch (args->mode & S_IFMT) {
@@ -1751,8 +1752,9 @@ linux_prctl(struct thread *td, struct linux_prctl_args *args)
 
 #ifdef DEBUG
 	if (ldebug(prctl))
-		printf(ARGS(prctl, "%d, %d, %d, %d, %d"), args->option,
-		    args->arg2, args->arg3, args->arg4, args->arg5);
+		printf(ARGS(prctl, "%d, %jd, %jd, %jd, %jd"), args->option,
+		    (intmax_t)args->arg2, (intmax_t)args->arg3,
+		    (intmax_t)args->arg4, (intmax_t)args->arg5);
 #endif
 
 	switch (args->option) {
